contact edit intregreated
