+++ src/App.js
@@ -1,6 +1,11 @@
-import React, { Component } from 'react'
-import { HashRouter, Route, Switch } from 'react-router-dom'
+import React, { Component, useState, useEffect } from 'react'
+import {BrowserRouter as Router, HashRouter, Route, Switch } from 'react-router-dom'
 import './scss/style.scss'
+import { uuid } from "uuidv4";
+import api from "./api/Contact";
+import './ContactComponents/App.css'
+
+
 
 const loading = (
   <div className="pt-3 text-center">
@@ -16,13 +21,100 @@ const Login = React.lazy(() => import('./views/pages/login/Login'))
 const Register = React.lazy(() => import('./views/pages/register/Register'))
 const Page404 = React.lazy(() => import('./views/pages/page404/Page404'))
 const Page500 = React.lazy(() => import('./views/pages/page500/Page500'))
+const Header =  React.lazy(() => import('./ContactComponents/Header')) 
+const AddContact = React.lazy(() => import('./ContactComponents/AddContact')) 
+const ContactList = React.lazy(() => import('./ContactComponents/ContactList')) 
+const ContactDetail = React.lazy(() => import('./ContactComponents/ContactDetail')) 
+const EditContact = React.lazy(() => import('./ContactComponents/EditContact')) 
 
-class App extends Component {
-  render() {
+function App() {
+
+    const LOCAL_STORAGE_KEY = "contacts";
+  const [contacts, setContacts] = useState([]);
+
+  //RetrieveContacts
+  const retrieveContacts = async () => {
+    const response = await api.get("/contacts");
+    return response.data;
+  };
+
+  const addContactHandler = async (contact) => {
+    console.log(contact);
+    const request = {
+      id: uuid(),
+      ...contact,
+    };
+
+    const response = await api.post("/contacts", request);
+    console.log(response);
+    setContacts([...contacts, response.data]);
+  };
+
+  const updateContactHandler = async (contact) => {
+    const response = await api.put(`/contacts/${contact.id}`, contact);
+    const { id, name, email } = response.data;
+    setContacts(
+      contacts.map((contact) => {
+        return contact.id === id ? { ...response.data } : contact;
+      })
+    );
+  };
+
+  const removeContactHandler = async (id) => {
+    await api.delete(`/contacts/${id}`);
+    const newContactList = contacts.filter((contact) => {
+      return contact.id !== id;
+    });
+
+    setContacts(newContactList);
+  };
+
+  useEffect(() => {
+    // const retriveContacts = JSON.parse(localStorage.getItem(LOCAL_STORAGE_KEY));
+    // if (retriveContacts) setContacts(retriveContacts);
+    const getAllCOntacts = async () => {
+      const allContacts = await retrieveContacts();
+      if (allContacts) setContacts(allContacts);
+    };
+
+    getAllCOntacts();
+  }, []);
+
+  useEffect(() => {
+    //localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(contacts));
+  }, [contacts]);
     return (
       <HashRouter>
         <React.Suspense fallback={loading}>
           <Switch>
+
+          <Route
+            exact path="/details"
+            
+            render={(props) => (
+              <ContactList
+                {...props}
+                contacts={contacts}
+                getContactId={removeContactHandler}
+              />
+            )}
+          />
+
+          <Route
+            path="/edit"
+            render={(props) => (
+              <EditContact
+                {...props}
+                updateContactHandler={updateContactHandler}
+              />
+            )}
+          />
+           <Route
+            path="/add"
+            render={(props) => <AddContact {...props} addContactHandler={addContactHandler} />}
+          />
+
+          <Route path="/contact/:id" component={ContactDetail} />
             <Route exact path="/login" name="Login Page" render={(props) => <Login {...props} />} />
             <Route
               exact
@@ -37,7 +129,6 @@ class App extends Component {
         </React.Suspense>
       </HashRouter>
     )
-  }
 }
 
 export default App
